requirements.txt
langchain
streamlit
openai==0.28
python-dotenv